requirements.txt
# MISC
pytest
pydantic

# MLX Alternatives for Windows
torch
transformers
accelerate
bitsandbytes
scipy
tqdm
numpy
huggingface_hub

# MeloTTS
g2p_en
cached_path
librosa
torchaudio

# Main
playsound==1.2.2
webrtcvad
sounddevice

# Web Interface
flask
flask-cors

# Add openai-whisper for Whisper STT
openai-whisper

# Add faster-whisper for optimized STT
faster-whisper
ctranslate2

# Optional dependencies for vision gesture control
opencv-python
mediapipe
pycaw
comtypes
screen_brightness_control
keyboard